+++ Torquescape/src/org/andork/torquescape/model/Player.java
@@ -1,43 +1,43 @@
 package org.andork.torquescape.model;
 
-import javax.vecmath.Point3d;
-import javax.vecmath.Vector3d;
+import javax.vecmath.Point3f;
+import javax.vecmath.Vector3f;
 
 public class Player
 {
 	final TriangleBasis	basis						= new TriangleBasis( );
 	
-	final Vector3d		temp						= new Vector3d( );
+	final Vector3f		temp						= new Vector3f( );
 	
-	final Point3d		location					= new Point3d( );
-	final Vector3d		basisForward				= new Vector3d( );
-	final Vector3d		basisUp						= new Vector3d( );
-	final Vector3d		cameraUp					= new Vector3d( );
-	final Vector3d		modelForward				= new Vector3d( );
-	final Vector3d		modelUp						= new Vector3d( );
+	final Point3f		location					= new Point3f( );
+	final Vector3f		basisForward				= new Vector3f( );
+	final Vector3f		basisUp						= new Vector3f( );
+	final Vector3f		cameraUp					= new Vector3f( );
+	final Vector3f		modelForward				= new Vector3f( );
+	final Vector3f		modelUp						= new Vector3f( );
 	
-	double				velocity					= 0;
-	double				forwardAcceleration			= 5;
-	double				maxForwardVelocity			= 50;
-	double				brakeDeceleration			= 30;
-	double				reverseAcceleration			= 5;
-	double				maxReverseVelocity			= 5;
-	double				naturalDeceleration			= 15;
+	float				velocity					= 0;
+	float				forwardAcceleration			= 5;
+	float				maxForwardVelocity			= 50;
+	float				brakeDeceleration			= 30;
+	float				reverseAcceleration			= 5;
+	float				maxReverseVelocity			= 5;
+	float				naturalDeceleration			= 15;
 	
-	double				angularVelocity				= 0;
-	double				maxAngularVelocity			= Math.PI / 2;
-	double				angularAcceleration			= Math.PI * 8;
-	double				naturalAngularDeceleration	= Math.PI * 8;
+	float				angularVelocity				= 0;
+	float				maxAngularVelocity			= ( float ) Math.PI / 2;
+	float				angularAcceleration			= ( float ) Math.PI * 8;
+	float				naturalAngularDeceleration	= ( float ) Math.PI * 8;
 	
 	public void setBasis( Triangle triangle )
 	{
 		basis.set( triangle );
 	}
 	
-	public void updateVelocity( double timestep , double forwardTime , double reverseTime , double leftTime , double rightTime )
+	public void updateVelocity( float timestep , float forwardTime , float reverseTime , float leftTime , float rightTime )
 	{
 		velocity += forwardAcceleration * forwardTime;
-		double reverseAmount;
+		float reverseAmount;
 		
 		if( velocity > 0 )
 		{
@@ -63,7 +63,7 @@ public class Player
 			velocity -= reverseAmount;
 		}
 		
-		double naturalDecelAmount = ( timestep - forwardTime - reverseTime ) * naturalDeceleration;
+		float naturalDecelAmount = ( timestep - forwardTime - reverseTime ) * naturalDeceleration;
 		if( naturalDecelAmount > Math.abs( velocity ) )
 		{
 			velocity = 0;
@@ -76,7 +76,7 @@ public class Player
 		velocity = Math.max( -maxReverseVelocity , Math.min( maxForwardVelocity , velocity ) );
 		
 		angularVelocity += angularAcceleration * ( leftTime - rightTime );
-		double angularDecelAmount = naturalAngularDeceleration * ( timestep - leftTime - rightTime );
+		float angularDecelAmount = naturalAngularDeceleration * ( timestep - leftTime - rightTime );
 		if( angularDecelAmount > Math.abs( angularVelocity ) )
 		{
 			angularVelocity = 0;
@@ -88,82 +88,82 @@ public class Player
 		angularVelocity = Math.signum( angularVelocity ) * Math.min( maxAngularVelocity , Math.abs( angularVelocity ) );
 	}
 	
-	public void getLocation( Point3d location )
+	public void getLocation( Point3f location )
 	{
 		location.set( this.location );
 	}
 	
-	public void getBasisForward( Vector3d forward )
+	public void getBasisForward( Vector3f forward )
 	{
 		forward.set( this.basisForward );
 	}
 	
-	public void getModelForward( Vector3d forward )
+	public void getModelForward( Vector3f forward )
 	{
 		forward.set( this.modelForward );
 	}
 	
-	public void getModelUp( Vector3d up )
+	public void getModelUp( Vector3f up )
 	{
 		up.set( this.modelUp );
 	}
 	
-	public void getCameraUp( Vector3d cameraUp )
+	public void getCameraUp( Vector3f cameraUp )
 	{
 		cameraUp.set( this.cameraUp );
 	}
 	
-	public void getBasisUp( Vector3d targetUp )
+	public void getBasisUp( Vector3f targetUp )
 	{
 		targetUp.set( this.basisUp );
 	}
 	
-	public double getVelocity( )
+	public float getVelocity( )
 	{
 		return velocity;
 	}
 	
-	public void setVelocity( double velocity )
+	public void setVelocity( float velocity )
 	{
 		this.velocity = velocity;
 	}
 	
-	public double getAngularVelocity( )
+	public float getAngularVelocity( )
 	{
 		return angularVelocity;
 	}
 	
-	public void setAngularVelocity( double angularVelocity )
+	public void setAngularVelocity( float angularVelocity )
 	{
 		this.angularVelocity = angularVelocity;
 	}
 	
-	public void setBasisForward( Vector3d forward )
+	public void setBasisForward( Vector3f forward )
 	{
 		this.basisForward.set( forward );
 	}
 	
-	public void setModelForward( Vector3d forward )
+	public void setModelForward( Vector3f forward )
 	{
 		this.modelForward.set( forward );
 	}
 	
-	public void setModelUp( Vector3d up )
+	public void setModelUp( Vector3f up )
 	{
 		this.modelUp.set( up );
 	}
 	
-	public void setCameraUp( Vector3d cameraUp )
+	public void setCameraUp( Vector3f cameraUp )
 	{
 		this.cameraUp.set( cameraUp );
 	}
 	
-	public void setBasisUp( Vector3d targetUp )
+	public void setBasisUp( Vector3f targetUp )
 	{
 		this.basisUp.set( targetUp );
 	}
 	
-	public void setLocation( Point3d location )
+	public void setLocation( Point3f location )
 	{
 		this.location.set( location );
 	}
